Added linear_search.cpp file.
I have added the linear_search.cpp file which shows how to implement linear search in c++ language. I It contains comments for all the steps which makes it beginner friendly and anyone can understand as well as implement the algorithm from this file. I have also explained the linear search algorithm in comments in the top of the file. This will definitely help all the students as well as anyone aspiring for placements.